requirements.txt
python-dotenv
pyupbit
matplotlib
numpy
pandas